Split collection indexing tests into their own module
apps/documents/tests/test_tasks.py had grown to cover zip creation,
document-source syncing, collection deletion, versioning, and collection
indexing in one file. Move the indexing block into
test_collection_indexing.py and lift the shared remote-index collection
fixture into a package conftest, renamed to remote_collection_index so it
does not shadow test_document_sources.py's own collection fixture, which
builds a materially different object.
